add an icon, a steering wheel obtained for free from: http://www.fastfreeimages.com/index.php/pages/images-steering_wheel.html
image terms are:


Terms Of Use


Thank you for taking a moment to review the terms and conditions under which you may make use of the content provided on this website.  We have done our best to craft a policy which allows you to make full use of our content for its intended purposes, while at the same time protecting our rights to it.

To clarify in plain English:
Permitted uses: Websites, newsletters, flyers, menus, signs, and other related materials, either business or personal, where the images are used to enhance the existing content.
Prohibited uses: Image or clipart collections, either online or archived, or any other means which offers the images for others to use; Logos; Any application where our images comprise most or all of the finished product.



 
The terms and conditions set forth in this document govern how you may use the content on this website. If you do not agree with any portion of this policy, you must destroy all downloaded content, any copies made, and any derivative works based upon the content.

1. The contents of the "free gallery", also referred to as "freebies", "free stuff", or "free images", are copyrighted works and are not in the public domain. They are being made available for you to use free of charge for your projects, provided that their use does not violate any portion of this usage policy.

2. No content on this website, free or paid, shall be published, archived, or otherwise provided in any way that offers it to others for their own use. This includes, but is not limited to, clipart or image websites, cd-rom or other offline collections, and file sharing networks.

3. Images on this website shall not be linked to from any other website. All images that you wish to use must be downloaded to your hard drive and stored on your own website space.

4. Any paid content on this website is to be used only on projects you or your company are directly participating in. Neither this website's contents nor your username or password may be shared with any other parties.

5. All content on this website shall at all times remain our property. It is being licensed to you for use as governed by the terms of this agreement. Should this agreement be breached, all said rights to usage of this content shall be revoked and all copies and derivative works must be immediately destroyed.

We reserve the right to make changes to this agreement at any time without notice of any kind.